backend/requirements.txt
asgiref==3.3.1
astroid==2.5
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
isort==5.7.0
joblib==1.0.1
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.8.6
pylint==2.7.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
wrapt==1.12.1
certifi=2020.12.5 
chardet=4.0.0 
idna=2.10
requests=2.25.1 
urllib3=1.26.4 
vk-api=11.9.3
